Update README.md
The first problem builds a number of independent functions which rise

moderately in complexity. Early functions deal with basic syntax and

iteration through Rust’s Slices like type `&[i32]`. Later functions

will require handling of references and dealing with modest ownership

issues that may involve borrowing via the reference creation operator

`&` or cloning data with that method supported.

The second problem builds a small command line program which apes the

behavior of the classic `wc` UNIX utility: it counts the characters,

words, and lines in an input file. This will give a chance to build a

command line program and deal with basic file input in Rust.